+++ src/control/reducer/set-errors.spec.ts
@@ -44,4 +44,19 @@ describe('form control setErrorsReducer', () => {
     const resultState = setErrorsReducer(INITIAL_STATE, new SetErrorsAction(FORM_CONTROL_ID, {}));
     expect(resultState).toBe(INITIAL_STATE);
   });
+
+  it('should keep async errors', () => {
+    const syncErrors = { required: true };
+    const asyncErrors = { $required: true };
+    const state = { ...INITIAL_STATE, isValid: false, isInvalid: true, errors: asyncErrors };
+    const resultState = setErrorsReducer(state, new SetErrorsAction(FORM_CONTROL_ID, syncErrors));
+    expect(resultState.errors).toEqual({ ...asyncErrors, ...syncErrors });
+  });
+
+  it('should throw if trying to set invalid error value', () => {
+    expect(() => setErrorsReducer(INITIAL_STATE, new SetErrorsAction(FORM_CONTROL_ID, null as any))).toThrowError();
+    expect(() => setErrorsReducer(INITIAL_STATE, new SetErrorsAction(FORM_CONTROL_ID, 1 as any))).toThrowError();
+    expect(() => setErrorsReducer(INITIAL_STATE, new SetErrorsAction(FORM_CONTROL_ID, [] as any))).toThrowError();
+    expect(() => setErrorsReducer(INITIAL_STATE, new SetErrorsAction(FORM_CONTROL_ID, { $required: true }))).toThrowError();
+  });
 });